Add files via upload
The 3 main files are the Register.php, Login.php, and Index.html. The Index.html is are main page to register, login into different role. The Register php is a function page to help the user create their own admin login, and store it in the database. And, the login.php help the user connect with the database and redirect the user into their right place.